Add comprehensive tests and extras for key provider plugin
