+++ pkgs/standards/swarmauri_keyproviders/tests/perf/test_ssh_create_key_perf.py
@@ -1,0 +1,21 @@
+import asyncio
+
+import pytest
+
+from swarmauri_keyproviders import SshKeyProvider
+from swarmauri_core.keys.types import KeySpec, KeyAlg, KeyClass, ExportPolicy, KeyUse
+
+
+@pytest.mark.perf
+def test_create_key_performance(benchmark) -> None:
+    spec = KeySpec(
+        klass=KeyClass.asymmetric,
+        alg=KeyAlg.ED25519,
+        uses=(KeyUse.SIGN, KeyUse.VERIFY),
+        export_policy=ExportPolicy.PUBLIC_ONLY,
+    )
+
+    def run() -> None:
+        asyncio.run(SshKeyProvider().create_key(spec))
+
+    benchmark(run)
