+++ pkgs/standards/swarmauri_keyproviders/tests/rfc/test_rfc8037_ed25519.py
@@ -1,0 +1,19 @@
+import pytest
+
+from swarmauri_keyproviders import SshKeyProvider
+from swarmauri_core.keys.types import KeySpec, KeyAlg, KeyClass, ExportPolicy, KeyUse
+
+
+@pytest.mark.rfc
+@pytest.mark.asyncio
+async def test_ed25519_jwk_matches_rfc8037() -> None:
+    kp = SshKeyProvider()
+    spec = KeySpec(
+        klass=KeyClass.asymmetric,
+        alg=KeyAlg.ED25519,
+        uses=(KeyUse.SIGN, KeyUse.VERIFY),
+        export_policy=ExportPolicy.PUBLIC_ONLY,
+    )
+    ref = await kp.create_key(spec)
+    jwk = await kp.get_public_jwk(ref.kid)
+    assert jwk["kty"] == "OKP" and jwk["crv"] == "Ed25519"
